backend/requirements.txt
Django>=4.2.7
pyjwt>=2.8.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.7